Naive bayes classifier (bi-class) from scratch using classes.
Implementation of naive bayes classifier on the tumor data set, which contains various tumor parameters as features and two classes : Benign and Malignant. Build the code from the scratch with the class named as naive bayes with various functions such as fit, predict, etc. The build up is done the concept of scikit learn library and its principle. Accuracy of 95% is achieved in this code.